recommend using git when using pass-grave
One of the easiest ways to ensure the data integrity of a password store
is to use the in-built git integration offered by pass. Since pass-grave
compresses, decompresses, and deletes data inside password-store, it is
essential that a user must ensure the data integrity of their
password-store in case of an unforseen event involving data loss.